+++ CHANGELOG.md
@@ -27,8 +27,8 @@
 - Restrict delivery partners so they can only view and edit their own organisation
 - Fund managers can manage users, organisations, funds, fund activites and fund transactions
 - Transaction and Activity dates are restricted to 10 years in the past or 25 years in the future at most
-- Fund managers can create basic programmes
 - Provide a way to flag an organisation as BEIS
 - User email addresses must be valid emails
 - Users are only associated with one organisation
 - Users land on their organisation#show page when they log in, instead of a "dashboard"
+- Fund manager can add a programme level activity